detect browser version and revert to default theme if you've < 69 This is because current FF release (67) theme's functions don't work - in particular: - browser.theme.getCurrent() always return an empty object - browser.theme.onUpdated listener never triggers
MDN indicate these are "fully working" on FF58 but I can't make them
work. Firefox nightly has no issue with this, so I'm guessing it's been
fixed recently.

Finally note that this means that on < 69 this extension will break any
custom theme (it will revert to default) since browser.theme.reset() is
also buggy on Firefox and does not reset to user default, but to the
Firefox's builtin default (that one is not fixed in any version)